Remove residual Multi-device Conversation (Conversation container) code
Removes the now-dead Conversation container and the ConversationTranslator internal machinery left behind after #1019, while preserving everything still shared with the live ConversationTranscriber, MeetingTranscriber, and DialogServiceConnector features.

- Delete the Conversation container and MDC-only public types (ConversationExpirationEventArgs, ConversationParticipantsChangedEventArgs, ConversationTranslation{Result,EventArgs,CanceledEventArgs}, ConversationInfo/IConversation, ParticipantChangedReason, ConversationHandler interfaces).

- Delete the internal ConversationTranslator recognizer chain (ConversationTranslatorRecognizer/ConversationRecognizerFactory, ConversationServiceAdapter, ConversationManager, ConversationConnection*, ConversationRequestSession, ConversationWebsocketMessageFormatter, ConversationTranslatorEventArgs, ConversationTranslatorInterfaces, ServiceMessages payloads).

- Move the still-shared IInternalParticipant/InternalParticipants into their own file (used by the live Meeting/MeetingTranscriber path).

- Rewrite Meeting to keep only the live transcriber-support members; drop the dead MDC host-flow and event handlers.

- Remove the dead conversation accessors from TranscriberRecognizer and TranscriptionServiceRecognizer (live code uses the meeting accessors); drop the dead ServiceRecognizerBase.conversationTranslatorToken setter and unused ConversationConnectionConfig members.

AB#32616
