+++ src/common.speech/Transcription/Exports.ts
@@ -1,21 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license.
 
-export { ConversationManager } from "./ConversationManager.js";
 export { ConversationConnectionConfig } from "./ConversationConnectionConfig.js";
-export { ConversationRecognizerFactory } from "./ConversationTranslatorRecognizer.js";
 export { TranscriberRecognizer } from "./TranscriberRecognizer.js";
 export {
-    ConversationReceivedTranslationEventArgs,
-    LockRoomEventArgs,
-    MuteAllEventArgs,
-    ParticipantAttributeEventArgs,
-    ParticipantEventArgs,
-    ParticipantsListEventArgs } from "./ConversationTranslatorEventArgs.js";
-export {
-    ConversationRecognizer,
-    ConversationTranslatorCommandTypes,
-    ConversationTranslatorMessageTypes,
-    IInternalConversation,
     IInternalParticipant,
-    InternalParticipants} from "./ConversationTranslatorInterfaces.js";
+    InternalParticipants } from "./InternalParticipants.js";
