--- src/common.speech/Transcription/ServiceMessages/Exports.ts
@@ -1,3 +1,0 @@
-export { CommandResponsePayload } from "./CommandResponsePayload.js";
-export { IParticipantsListPayloadResponse, IParticipantPayloadResponse, ParticipantsListPayloadResponse, ParticipantPayloadResponse } from "./ParticipantResponsePayload.js";
-export { ITranslationResponsePayload, SpeechResponsePayload, TextResponsePayload } from "./TranslationResponsePayload.js";
