+++ src/sdk/Transcription/Exports.ts
@@ -1,18 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license.
-// Multi-device Conversation is a Preview feature.
 
-export { Conversation, ConversationImpl } from "./Conversation.js";
-export { ConversationInfo } from "./IConversation.js";
-export { ConversationExpirationEventArgs } from "./ConversationExpirationEventArgs.js";
-export { ConversationParticipantsChangedEventArgs } from "./ConversationParticipantsChangedEventArgs.js";
-export { ConversationTranslationCanceledEventArgs } from "./ConversationTranslationCanceledEventArgs.js";
-export { ConversationTranslationEventArgs } from "./ConversationTranslationEventArgs.js";
-export { ConversationTranslationResult } from "./ConversationTranslationResult.js";
 export { ConversationTranscriber } from "./ConversationTranscriber.js";
 export { IParticipant, Participant, User, VoiceSignature } from "./IParticipant.js";
-export { ParticipantChangedReason } from "./ParticipantChangedReason.js";
-export { ConversationHandler, ConversationTranscriptionHandler, IConversationServiceHandler } from "./ConversationHandler.js";
 export { Meeting, MeetingImpl } from "./Meeting.js";
 export { MeetingInfo } from "./IMeeting.js";
 export { MeetingTranscriptionCanceledEventArgs } from "./MeetingTranscriptionCanceledEventArgs.js";
